+++ SDWebImage/SDWebImageDownloaderOperation.m
@@ -332,6 +332,7 @@ didReceiveResponse:(NSURLResponse *)response
                 if ([coder conformsToProtocol:@protocol(SDWebImageProgressiveCoder)] &&
                     [((id<SDWebImageProgressiveCoder>)coder) canIncrementallyDecodeFromData:imageData]) {
                     self.progressiveCoder = [[[coder class] alloc] init];
+                    break;
                 }
             }
         }