+++ SDWebImage/UIImage+MultiFormat.m
@@ -13,6 +13,31 @@
 
 @implementation UIImage (MultiFormat)
 
+#if SD_MAC
+- (NSUInteger)sd_imageLoopCount {
+    NSUInteger imageLoopCount = 0;
+    for (NSImageRep *rep in self.representations) {
+        if ([rep isKindOfClass:[NSBitmapImageRep class]]) {
+            NSBitmapImageRep *bitmapRep = (NSBitmapImageRep *)rep;
+            imageLoopCount = [[bitmapRep valueForProperty:NSImageLoopCount] unsignedIntegerValue];
+            break;
+        }
+    }
+    return imageLoopCount;
+}
+
+- (void)setSd_imageLoopCount:(NSUInteger)sd_imageLoopCount {
+    for (NSImageRep *rep in self.representations) {
+        if ([rep isKindOfClass:[NSBitmapImageRep class]]) {
+            NSBitmapImageRep *bitmapRep = (NSBitmapImageRep *)rep;
+            [bitmapRep setProperty:NSImageLoopCount withValue:@(sd_imageLoopCount)];
+            break;
+        }
+    }
+}
+
+#else
+
 - (NSUInteger)sd_imageLoopCount {
     NSUInteger imageLoopCount = 0;
     NSNumber *value = objc_getAssociatedObject(self, @selector(sd_imageLoopCount));
@@ -26,6 +51,7 @@
     NSNumber *value = @(sd_imageLoopCount);
     objc_setAssociatedObject(self, @selector(sd_imageLoopCount), value, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
 }
+#endif
 
 + (nullable UIImage *)sd_imageWithData:(nullable NSData *)data {
     return [[SDWebImageCodersManager sharedInstance] decodedImageWithData:data];